Separating out click functionality, added jshint dependency, and linted code
